remove playground crate. move root crate to rippled-bridge crate. move some code from rippled-bridge lib.rs to sub-modules.
