+++ rust/crates/rippled-bridge/src/lib.rs
@@ -4,10 +4,14 @@ use cxx::{ExternType, SharedPtr, type_id, UniquePtr};
 use cxx::kind::Trivial;
 
 pub mod dummy_tx_rs;
+pub mod ter;
+pub mod flags;
 
 pub use dummy_tx_rs::pre_flight;
 pub use dummy_tx_rs::pre_claim;
 pub use dummy_tx_rs::do_apply;
+pub use ter::{TER, NotTEC, TEFcodes, TEMcodes, TELcodes, TECcodes, TEScodes, TERcodes};
+pub use flags::{LedgerSpecificFlags, ApplyFlags};
 
 // Also try this
 #[no_mangle]
@@ -28,203 +32,6 @@ pub fn calculateBaseFee(view: &rippled::ReadView, tx: &rippled::STTx) -> XRPAmou
 #[no_mangle]
 pub fn doApply(mut ctx: Pin<&mut rippled::ApplyContext>, mPriorBalance: rippled::XRPAmount, mSourceBalance: rippled::XRPAmount) -> TER {
     do_apply(ctx, mPriorBalance, mSourceBalance)
-}
-
-#[repr(i32)]
-#[derive(Clone, Copy)]
-pub enum TELcodes {
-    telLOCAL_ERROR = -399,
-    telBAD_DOMAIN,
-    telBAD_PATH_COUNT,
-    telBAD_PUBLIC_KEY,
-    telFAILED_PROCESSING,
-    telINSUF_FEE_P,
-    telNO_DST_PARTIAL,
-    telCAN_NOT_QUEUE,
-    telCAN_NOT_QUEUE_BALANCE,
-    telCAN_NOT_QUEUE_BLOCKS,
-    telCAN_NOT_QUEUE_BLOCKED,
-    telCAN_NOT_QUEUE_FEE,
-    telCAN_NOT_QUEUE_FULL,
-}
-
-unsafe impl cxx::ExternType for TELcodes {
-    type Id = type_id!("ripple::TELcodes");
-    type Kind = Trivial;
-}
-
-#[repr(i32)]
-#[derive(Clone, Copy)]
-pub enum TEMcodes {
-    temMALFORMED = -299,
-
-    temBAD_AMOUNT,
-    temBAD_CURRENCY,
-    temBAD_EXPIRATION,
-    temBAD_FEE,
-    temBAD_ISSUER,
-    temBAD_LIMIT,
-    temBAD_OFFER,
-    temBAD_PATH,
-    temBAD_PATH_LOOP,
-    temBAD_REGKEY,
-    temBAD_SEND_XRP_LIMIT,
-    temBAD_SEND_XRP_MAX,
-    temBAD_SEND_XRP_NO_DIRECT,
-    temBAD_SEND_XRP_PARTIAL,
-    temBAD_SEND_XRP_PATHS,
-    temBAD_SEQUENCE,
-    temBAD_SIGNATURE,
-    temBAD_SRC_ACCOUNT,
-    temBAD_TRANSFER_RATE,
-    temDST_IS_SRC,
-    temDST_NEEDED,
-    temINVALID,
-    temINVALID_FLAG,
-    temREDUNDANT,
-    temRIPPLE_EMPTY,
-    temDISABLED,
-    temBAD_SIGNER,
-    temBAD_QUORUM,
-    temBAD_WEIGHT,
-    temBAD_TICK_SIZE,
-    temINVALID_ACCOUNT_ID,
-    temCANNOT_PREAUTH_SELF,
-    temINVALID_COUNT,
-
-    temUNCERTAIN,  // An internal intermediate result; should never be returned.
-    temUNKNOWN,    // An internal intermediate result; should never be returned.
-
-    temSEQ_AND_TICKET,
-    temBAD_NFTOKEN_TRANSFER_FEE,
-}
-
-unsafe impl cxx::ExternType for TEMcodes {
-    type Id = type_id!("ripple::TEMcodes");
-    type Kind = Trivial;
-}
-
-#[repr(i32)]
-#[derive(Clone, Copy)]
-pub enum TEFcodes {
-    tefFAILURE = -199,
-    tefALREADY,
-    tefBAD_ADD_AUTH,
-    tefBAD_AUTH,
-    tefBAD_LEDGER,
-    tefCREATED,
-    tefEXCEPTION,
-    tefINTERNAL,
-    tefNO_AUTH_REQUIRED,
-    // Can't set auth if auth is not required.
-    tefPAST_SEQ,
-    tefWRONG_PRIOR,
-    tefMASTER_DISABLED,
-    tefMAX_LEDGER,
-    tefBAD_SIGNATURE,
-    tefBAD_QUORUM,
-    tefNOT_MULTI_SIGNING,
-    tefBAD_AUTH_MASTER,
-    tefINVARIANT_FAILED,
-    tefTOO_BIG,
-    tefNO_TICKET,
-    tefNFTOKEN_IS_NOT_TRANSFERABLE,
-}
-
-unsafe impl cxx::ExternType for TEFcodes {
-    type Id = type_id!("ripple::TEFcodes");
-    type Kind = Trivial;
-}
-
-#[repr(i32)]
-#[derive(Clone, Copy)]
-pub enum TERcodes {
-    terRETRY = -99,
-    terFUNDS_SPENT,  // DEPRECATED.
-    terINSUF_FEE_B,  // Can't pay fee, therefore don't burden network.
-    terNO_ACCOUNT,   // Can't pay fee, therefore don't burden network.
-    terNO_AUTH,      // Not authorized to hold IOUs.
-    terNO_LINE,      // Internal flag.
-    terOWNERS,       // Can't succeed with non-zero owner count.
-    terPRE_SEQ,      // Can't pay fee, no point in forwarding, so don't
-    // burden network.
-    terLAST,         // DEPRECATED.
-    terNO_RIPPLE,    // Rippling not allowed
-    terQUEUED,       // Transaction is being held in TxQ until fee drops
-    terPRE_TICKET,   // Ticket is not yet in ledger but might be on its way
-}
-
-unsafe impl cxx::ExternType for TERcodes {
-    type Id = type_id!("ripple::TERcodes");
-    type Kind = Trivial;
-}
-
-#[repr(i32)]
-#[derive(Clone, Copy)]
-pub enum TEScodes {
-    tesSUCCESS = 0
-}
-
-unsafe impl cxx::ExternType for TEScodes {
-    type Id = type_id!("ripple::TEScodes");
-    type Kind = Trivial;
-}
-
-#[repr(i32)]
-#[derive(Clone, Copy)]
-pub enum TECcodes {
-    tecCLAIM = 100,
-    tecPATH_PARTIAL = 101,
-    tecUNFUNDED_ADD = 102,  // Unused legacy code
-    tecUNFUNDED_OFFER = 103,
-    tecUNFUNDED_PAYMENT = 104,
-    tecFAILED_PROCESSING = 105,
-    tecDIR_FULL = 121,
-    tecINSUF_RESERVE_LINE = 122,
-    tecINSUF_RESERVE_OFFER = 123,
-    tecNO_DST = 124,
-    tecNO_DST_INSUF_XRP = 125,
-    tecNO_LINE_INSUF_RESERVE = 126,
-    tecNO_LINE_REDUNDANT = 127,
-    tecPATH_DRY = 128,
-    tecUNFUNDED = 129,
-    tecNO_ALTERNATIVE_KEY = 130,
-    tecNO_REGULAR_KEY = 131,
-    tecOWNERS = 132,
-    tecNO_ISSUER = 133,
-    tecNO_AUTH = 134,
-    tecNO_LINE = 135,
-    tecINSUFF_FEE = 136,
-    tecFROZEN = 137,
-    tecNO_TARGET = 138,
-    tecNO_PERMISSION = 139,
-    tecNO_ENTRY = 140,
-    tecINSUFFICIENT_RESERVE = 141,
-    tecNEED_MASTER_KEY = 142,
-    tecDST_TAG_NEEDED = 143,
-    tecINTERNAL = 144,
-    tecOVERSIZE = 145,
-    tecCRYPTOCONDITION_ERROR = 146,
-    tecINVARIANT_FAILED = 147,
-    tecEXPIRED = 148,
-    tecDUPLICATE = 149,
-    tecKILLED = 150,
-    tecHAS_OBLIGATIONS = 151,
-    tecTOO_SOON = 152,
-    tecHOOK_ERROR = 153,
-    tecMAX_SEQUENCE_REACHED = 154,
-    tecNO_SUITABLE_NFTOKEN_PAGE = 155,
-    tecNFTOKEN_BUY_SELL_MISMATCH = 156,
-    tecNFTOKEN_OFFER_TYPE_MISMATCH = 157,
-    tecCANT_ACCEPT_OWN_NFTOKEN_OFFER = 158,
-    tecINSUFFICIENT_FUNDS = 159,
-    tecOBJECT_NOT_FOUND = 160,
-    tecINSUFFICIENT_PAYMENT = 161,
-}
-
-unsafe impl cxx::ExternType for TECcodes {
-    type Id = type_id!("ripple::TECcodes");
-    type Kind = Trivial;
 }
 
 #[repr(u16)]
@@ -530,190 +337,6 @@ unsafe impl cxx::ExternType for XRPAmount {
     type Kind = Trivial;
 }
 
-/////////////////////////////
-// NotTEC
-/////////////////////////////
-
-#[repr(C)]
-#[derive(PartialEq)]
-pub struct NotTEC {
-    code: i32
-}
-
-unsafe impl cxx::ExternType for NotTEC {
-    type Id = type_id!("ripple::NotTEC");
-    type Kind = Trivial;
-}
-
-impl NotTEC {
-    pub fn new(code: i32) -> Self {
-        NotTEC { code }
-    }
-}
-
-impl From<TEScodes> for NotTEC {
-    fn from(value: TEScodes) -> Self {
-        NotTEC::new(value as i32)
-    }
-}
-
-impl From<TELcodes> for NotTEC {
-    fn from(value: TELcodes) -> Self {
-        NotTEC::new(value as i32)
-    }
-}
-
-
-impl From<TEMcodes> for NotTEC {
-    fn from(value: TEMcodes) -> Self {
-        NotTEC::new(value as i32)
-    }
-}
-
-impl From<TEFcodes> for NotTEC {
-    fn from(value: TEFcodes) -> Self {
-        NotTEC::new(value as i32)
-    }
-}
-
-impl From<TERcodes> for NotTEC {
-    fn from(value: TERcodes) -> Self {
-        NotTEC::new(value as i32)
-    }
-}
-
-impl PartialEq<TEScodes> for NotTEC {
-    fn eq(&self, other: &TEScodes) -> bool {
-        self.code == (*other as i32)
-    }
-}
-
-impl PartialEq<TELcodes> for NotTEC {
-    fn eq(&self, other: &TELcodes) -> bool {
-        self.code == (*other as i32)
-    }
-}
-
-impl PartialEq<TEMcodes> for NotTEC {
-    fn eq(&self, other: &TEMcodes) -> bool {
-        self.code == (*other as i32)
-    }
-}
-
-impl PartialEq<TEFcodes> for NotTEC {
-    fn eq(&self, other: &TEFcodes) -> bool {
-        self.code == (*other as i32)
-    }
-}
-
-impl PartialEq<TERcodes> for NotTEC {
-    fn eq(&self, other: &TERcodes) -> bool {
-        self.code == (*other as i32)
-    }
-}
-
-impl PartialEq<TECcodes> for NotTEC {
-    fn eq(&self, other: &TECcodes) -> bool {
-        self.code == (*other as i32)
-    }
-}
-
-
-/////////////////////////////
-// NotTEC
-/////////////////////////////
-
-#[repr(C)]
-#[derive(PartialEq)]
-pub struct TER {
-    code: i32
-}
-
-unsafe impl cxx::ExternType for TER {
-    type Id = type_id!("ripple::TER");
-    type Kind = Trivial;
-}
-
-impl TER {
-    pub fn new(code: i32) -> Self {
-        TER { code }
-    }
-}
-
-impl From<TEScodes> for TER {
-    fn from(value: TEScodes) -> Self {
-        TER::new(value as i32)
-    }
-}
-
-impl From<TELcodes> for TER {
-    fn from(value: TELcodes) -> Self {
-        TER::new(value as i32)
-    }
-}
-
-
-impl From<TEMcodes> for TER {
-    fn from(value: TEMcodes) -> Self {
-        TER::new(value as i32)
-    }
-}
-
-impl From<TEFcodes> for TER {
-    fn from(value: TEFcodes) -> Self {
-        TER::new(value as i32)
-    }
-}
-
-impl From<TERcodes> for TER {
-    fn from(value: TERcodes) -> Self {
-        TER::new(value as i32)
-    }
-}
-
-impl From<TECcodes> for TER {
-    fn from(value: TECcodes) -> Self {
-        TER::new(value as i32)
-    }
-}
-
-///////////////////
-impl PartialEq<TEScodes> for TER {
-    fn eq(&self, other: &TEScodes) -> bool {
-        self.code == (*other as i32)
-    }
-}
-
-impl PartialEq<TELcodes> for TER {
-    fn eq(&self, other: &TELcodes) -> bool {
-        self.code == (*other as i32)
-    }
-}
-
-impl PartialEq<TEMcodes> for TER {
-    fn eq(&self, other: &TEMcodes) -> bool {
-        self.code == (*other as i32)
-    }
-}
-
-impl PartialEq<TEFcodes> for TER {
-    fn eq(&self, other: &TEFcodes) -> bool {
-        self.code == (*other as i32)
-    }
-}
-
-impl PartialEq<TERcodes> for TER {
-    fn eq(&self, other: &TERcodes) -> bool {
-        self.code == (*other as i32)
-    }
-}
-
-impl PartialEq<TECcodes> for TER {
-    fn eq(&self, other: &TECcodes) -> bool {
-        self.code == (*other as i32)
-    }
-}
-
 #[repr(i16)]
 #[derive(Clone, Copy)]
 pub enum LedgerEntryType {
@@ -867,105 +490,4 @@ pub struct Keylet {
 unsafe impl cxx::ExternType for Keylet {
     type Id = type_id!("ripple::Keylet");
     type Kind = Trivial;
-}
-
-#[repr(i32)]
-#[derive(Copy, Clone)]
-pub enum ApplyFlags {
-    tapNONE = 0x00,
-
-    // This is a local transaction with the
-    // fail_hard flag set.
-    tapFAIL_HARD = 0x10,
-
-    // This is not the transaction's last pass
-    // Transaction can be retried, soft failures allowed
-    tapRETRY = 0x20,
-
-    // Transaction came from a privileged source
-    tapUNLIMITED = 0x400,
-}
-
-unsafe impl cxx::ExternType for ApplyFlags {
-    type Id = type_id!("ripple::ApplyFlags");
-    type Kind = Trivial;
-}
-
-
-pub enum LedgerSpecificFlags {
-    // ltACCOUNT_ROOT
-    lsfPasswordSpent,  // True, if password set fee is spent.
-    lsfRequireDestTag,  // True, to require a DestinationTag for payments.
-    lsfRequireAuth,  // True, to require a authorization to hold IOUs.
-    lsfDisallowXRP,    // True, to disallow sending XRP.
-    lsfDisableMaster,  // True, force regular key
-    lsfNoFreeze,       // True, cannot freeze ripple states
-    lsfGlobalFreeze,   // True, all assets frozen
-    lsfDefaultRipple,               // True, trust lines allow rippling by default
-    lsfDepositAuth,  // True, all deposits require authorization
-    /*  // reserved for Hooks amendment
-        lsfTshCollect = 0x02000000,     // True, allow TSH collect-calls to acc hooks
-    */
-    lsfDisallowIncomingNFTokenOffer,// True, reject new incoming NFT offers
-    lsfDisallowIncomingCheck,               // True, reject new checks
-    lsfDisallowIncomingPayChan,               // True, reject new paychans
-    lsfDisallowIncomingTrustline,               // True, reject new trustlines (only if no issued assets)
-
-    // ltOFFER
-    lsfPassive,
-    lsfSell,  // True, offer was placed as a sell.
-
-    // ltRIPPLE_STATE
-    lsfLowReserve,  // True, if entry counts toward reserve.
-    lsfHighReserve,
-    lsfLowAuth,
-    lsfHighAuth,
-    lsfLowNoRipple,
-    lsfHighNoRipple,
-    lsfLowFreeze,   // True, low side has set freeze flag
-    lsfHighFreeze,  // True, high side has set freeze flag
-
-    // ltSIGNER_LIST
-    lsfOneOwnerCount,  // True, uses only one OwnerCount
-
-    // ltDIR_NODE
-    lsfNFTokenBuyOffers,
-    lsfNFTokenSellOffers,
-
-    // ltNFTOKEN_OFFER
-    lsfSellNFToken,
-}
-
-impl From<LedgerSpecificFlags> for u32 {
-    fn from(value: LedgerSpecificFlags) -> Self {
-        match value {
-            LedgerSpecificFlags::lsfPasswordSpent => 0x00010000,
-            LedgerSpecificFlags::lsfRequireDestTag => 0x00020000,
-            LedgerSpecificFlags::lsfRequireAuth => 0x00040000,
-            LedgerSpecificFlags::lsfDisallowXRP => 0x00080000,
-            LedgerSpecificFlags::lsfDisableMaster => 0x00100000,
-            LedgerSpecificFlags::lsfNoFreeze => 0x00200000,
-            LedgerSpecificFlags::lsfGlobalFreeze => 0x00400000,
-            LedgerSpecificFlags::lsfDefaultRipple => 0x00800000,
-            LedgerSpecificFlags::lsfDepositAuth => 0x01000000,
-            LedgerSpecificFlags::lsfDisallowIncomingNFTokenOffer => 0x04000000,
-            LedgerSpecificFlags::lsfDisallowIncomingCheck => 0x08000000,
-            LedgerSpecificFlags::lsfDisallowIncomingPayChan => 0x10000000,
-            LedgerSpecificFlags::lsfDisallowIncomingTrustline => 0x20000000,
-            LedgerSpecificFlags::lsfPassive => 0x00010000,
-            LedgerSpecificFlags::lsfSell => 0x00020000,
-            LedgerSpecificFlags::lsfLowReserve => 0x00010000,
-            LedgerSpecificFlags::lsfHighReserve => 0x00020000,
-            LedgerSpecificFlags::lsfLowAuth => 0x00040000,
-            LedgerSpecificFlags::lsfHighAuth => 0x00080000,
-            LedgerSpecificFlags::lsfLowNoRipple => 0x00100000,
-            LedgerSpecificFlags::lsfHighNoRipple => 0x00200000,
-            LedgerSpecificFlags::lsfLowFreeze => 0x00400000,
-            LedgerSpecificFlags::lsfHighFreeze => 0x00800000,
-            LedgerSpecificFlags::lsfOneOwnerCount => 0x00010000,
-            LedgerSpecificFlags::lsfNFTokenBuyOffers => 0x00000001,
-            LedgerSpecificFlags::lsfNFTokenSellOffers => 0x00000002,
-            LedgerSpecificFlags::lsfSellNFToken => 0x00000001,
-        }
-    }
 }
