--- rust/playground/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "playground"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-rust = { path = "../" }
-cxx = "1.0.94"
-"xrpl-rust-sdk-core" = { path = "../../../xrpl-rust-sdk/xrpl-rust-sdk-core" }
